Drop foreign key for sequence
When updating twice the same invoice, PostgreSQL takes a shared lock on the
foreign key row of ir.sequence.strict. This prevent the sequence to be used for
numbering during all the transaction.
As some processes may write twice on an invoice due to modularity and as they
can last too long, it is better to avoid such lock by removing the foreign key
constraint. Storing the sequence id on the invoice is only to ensure sequential
numbering not the integrity of the sequence record.

issue9822
review331161002