Convert report into bytes before storing
If the invoice report has a text output, we must encode it as utf-8 before
storing it.
And we do not need to convert to bytes the data read because Binary field is
already doing that.

issue10937
review376581002
(grafted from a22984e8d5324c2ee018fa9b91e89bb5cefefd4c)

--HG--
branch : 6.0